+++ pkg/scaling/resolver/scale_resolvers.go
@@ -288,8 +288,8 @@ func resolveAuthRef(ctx context.Context, client client.Client, logger logr.Logge
 				}
 			}
 			if triggerAuthSpec.GCPSecretManager != nil && len(triggerAuthSpec.GCPSecretManager.Secrets) > 0 {
-				vaultHandler := NewGCPSecretManagerHandler(ctx, triggerAuthSpec.GCPSecretManager)
-				err := vaultHandler.Initialize(client, logger, triggerNamespace, secretsLister)
+				vaultHandler := NewGCPSecretManagerHandler(triggerAuthSpec.GCPSecretManager)
+				err := vaultHandler.Initialize(ctx, client, logger, triggerNamespace, secretsLister)
 				if err != nil {
 					logger.Error(err, "error authenticating to GCP Secret Manager", "triggerAuthRef.Name", triggerAuthRef.Name)
 				} else {
@@ -298,7 +298,7 @@ func resolveAuthRef(ctx context.Context, client client.Client, logger logr.Logge
 						if secret.Version != "" {
 							version = secret.Version
 						}
-						res, err := vaultHandler.Read(secret.ID, version)
+						res, err := vaultHandler.Read(ctx, secret.ID, version)
 						if err != nil {
 							logger.Error(err, "error trying to read secret from GCP Secret Manager", "triggerAuthRef.Name", triggerAuthRef.Name,
 								"secret.Name", secret.ID, "secret.Version", secret.Version)